+++ lib/src/client_browser.dart
@@ -40,8 +40,8 @@ class ClientBrowser extends ClientBase with ClientMixin {
       'x-sdk-name': 'Flutter',
       'x-sdk-platform': 'client',
       'x-sdk-language': 'flutter',
-      'x-sdk-version': '23.1.0',
-      'X-Appwrite-Response-Format': '1.9.1',
+      'x-sdk-version': '24.0.0',
+      'X-Appwrite-Response-Format': '1.9.2',
     };
 
     config = {};