+++ test/src/models/membership_test.dart
@@ -11,6 +11,7 @@ void main() {
         userId: '5e5ea5c16897e',
         userName: 'John Doe',
         userEmail: 'john@appwrite.io',
+        userPhone: '+1 555 555 5555',
         teamId: '5e5ea5c16897e',
         teamName: 'VIP',
         invited: '2020-10-15T06:38:00.000+00:00',
@@ -29,6 +30,7 @@ void main() {
       expect(result.userId, '5e5ea5c16897e');
       expect(result.userName, 'John Doe');
       expect(result.userEmail, 'john@appwrite.io');
+      expect(result.userPhone, '+1 555 555 5555');
       expect(result.teamId, '5e5ea5c16897e');
       expect(result.teamName, 'VIP');
       expect(result.invited, '2020-10-15T06:38:00.000+00:00');